Changed the way of retrieving query metrics (getting results directly from query execution plan)

+++ GraphMigrator/GraphMigrator.Domain/Models/Neo4jPerformanceMetrics.cs
@@ -5,7 +5,7 @@ public class Neo4jPerformanceMetrics
     public long ExecutionTimeMs { get; set; }
     public long MemoryUsageBytes { get; set; }
     public double CpuPercentage { get; set; }
-    public int DatabaseHits { get; set; }
+    public long DatabaseHits { get; set; }
     public Dictionary<string, object> ProfileInfo { get; set; }
 }
